requirements.txt
streamlit
pandas
plotly
numpy
mysql-connector-python